move build to webpack
