+++ index.html
@@ -42,6 +42,11 @@
         </div>
       </div>
     </div>
-    <script src="main.js"></script>
+    <script src="dist/app.bundle.js"></script>
+    <script>
+      document.addEventListener('DOMContentLoaded', function() {
+        MagounStatusBoardUI.init();
+      });
+    </script>
   </body>
 </html>
