+++ package.json
@@ -4,7 +4,8 @@
   "description": "Magoun status board UI",
   "main": "main.js",
   "scripts": {
-    "serve": "http-server ."
+    "serve": "http-server .",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
@@ -17,6 +18,13 @@
   },
   "homepage": "https://github.com/bencentra/MagounStatusBoardUI#readme",
   "devDependencies": {
-    "http-server": "^0.10.0"
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
+    "http-server": "^0.10.0",
+    "webpack": "^3.8.1"
+  },
+  "dependencies": {
+    "babel-polyfill": "^6.26.0"
   }
 }
