+++ src/fetch.js
@@ -1,0 +1,9 @@
+async function fetchJSON(url) {
+  const response = await fetch(url);
+  const json = await response.json();
+  return json;
+}
+
+export {
+  fetchJSON,
+};
