+++ webpack.config.js
@@ -1,0 +1,20 @@
+const path = require('path');
+
+module.exports = {
+  entry: ['babel-polyfill', './src/index.js'],
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'app.bundle.js',
+    library: 'MagounStatusBoardUI',
+    libraryTarget: 'umd',
+  },
+  devtool: 'inline-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        use: 'babel-loader',
+      }
+    ]
+  }
+};
